test: move upsert_peer_session tests beside their implementation
