+++ rust/src/mostro/session.rs
@@ -230,3 +230,105 @@ static SESSION_MGR: OnceLock<SessionManager> = OnceLock::new();
 pub fn session_manager() -> &'static SessionManager {
     SESSION_MGR.get_or_init(SessionManager::new)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::api::types::{OrderKind, OrderStatus, SellerStep, TradeStep};
+
+    fn dummy_order_info(id: &str) -> OrderInfo {
+        OrderInfo {
+            id: id.to_string(),
+            kind: OrderKind::Buy,
+            status: OrderStatus::Pending,
+            fiat_code: "USD".to_string(),
+            fiat_amount: Some(100.0),
+            fiat_amount_min: None,
+            fiat_amount_max: None,
+            payment_method: "Bank".to_string(),
+            premium: 0.0,
+            is_mine: false,
+            created_at: 0,
+            expires_at: None,
+            amount_sats: None,
+            creator_pubkey: String::new(),
+            rating: 0.0,
+            total_reviews: 0,
+            days_active: 0,
+        }
+    }
+
+    fn dummy_trade_info(order_id: &str, role: TradeRole, trade_key_index: u32) -> TradeInfo {
+        TradeInfo {
+            id: uuid::Uuid::new_v4().to_string(),
+            order: dummy_order_info(order_id),
+            role,
+            counterparty_pubkey: String::new(),
+            current_step: TradeStep::Seller(SellerStep::TakerFound),
+            hold_invoice: None,
+            buyer_invoice: None,
+            trade_key_index,
+            cooperative_cancel_state: None,
+            timeout_at: None,
+            started_at: 0,
+            completed_at: None,
+            outcome: None,
+            peer_rating: None,
+            peer_reviews: None,
+            peer_days: None,
+            rated_at: None,
+        }
+    }
+
+    /// ermeme review, PR #347: the maker never passes through `take_order`'s
+    /// create_session, so when the peer reveal arrives there is no session to
+    /// update. The reveal handler (and startup chat recovery) must be able to
+    /// materialize one from the persisted trade row, or `send_message` stays
+    /// on its silent local-only path and the peer never receives anything.
+    #[tokio::test]
+    async fn maker_reveal_without_a_session_builds_one_from_the_trade_row() {
+        let order_id = uuid::Uuid::new_v4().to_string();
+        let trade = dummy_trade_info(&order_id, TradeRole::Seller, 42);
+        let peer_hex = "00000000000000000000000000000000000000000000000000000000000000aa";
+
+        let mgr = session_manager();
+        assert!(mgr.get_session(&order_id).await.is_none());
+
+        mgr.upsert_peer_session(&trade, peer_hex, [7u8; 32]).await;
+
+        let session = mgr
+            .get_session(&order_id)
+            .await
+            .expect("session materialized from the trade row");
+        assert_eq!(session.peer_pubkey.as_deref(), Some(peer_hex));
+        assert_eq!(session.shared_key, Some([7u8; 32]));
+        assert_eq!(session.trade_key_index, 42);
+        assert!(matches!(session.role, TradeRole::Seller));
+        assert_eq!(session.order.id, order_id);
+    }
+
+    /// A live session (taker path, or a daemon replay) must survive the
+    /// upsert untouched except for the peer material — and unlike
+    /// create_session, the upsert must not fail on the existing entry.
+    #[tokio::test]
+    async fn upsert_peer_session_only_touches_peer_material_on_an_existing_session() {
+        let order_id = uuid::Uuid::new_v4().to_string();
+        let order = dummy_order_info(&order_id);
+
+        let mgr = session_manager();
+        mgr.create_session(order_id.clone(), TradeRole::Buyer, 7, order)
+            .await
+            .expect("live session");
+
+        // The trade row disagrees on role and index: the live values must win.
+        let trade = dummy_trade_info(&order_id, TradeRole::Seller, 99);
+        let peer_hex = "00000000000000000000000000000000000000000000000000000000000000bb";
+        mgr.upsert_peer_session(&trade, peer_hex, [9u8; 32]).await;
+
+        let session = mgr.get_session(&order_id).await.expect("still one session");
+        assert_eq!(session.peer_pubkey.as_deref(), Some(peer_hex));
+        assert_eq!(session.shared_key, Some([9u8; 32]));
+        assert_eq!(session.trade_key_index, 7);
+        assert!(matches!(session.role, TradeRole::Buyer));
+    }
+}
